fix to python3

+++ Flask_By_Example/chapter11/waitercaller.py
@@ -1,34 +1,18 @@
 import datetime
+from flask import Flask, redirect, render_template, request, url_for
+from flask_login import LoginManager, current_user, login_required, login_user, logout_user
 
 import config
-from flask import Flask
-from flask import redirect
-from flask import render_template
-from flask import request
-from flask import url_for
-from flask.ext.login import LoginManager
-from flask.ext.login import current_user
-from flask.ext.login import login_required
-from flask.ext.login import login_user
-from flask.ext.login import logout_user
-
-if config.test:
-    from mockdbhelper import MockDBHelper as DBHelper
-else:
-    from dbhelper import DBHelper
-
+from mockdbhelper import MockDBHelper as DBHelper if config.test else DBHelper
 from passwordhelper import PasswordHelper
 from bitlyhelper import BitlyHelper
 from user import User
-
-from forms import RegistrationForm
-from forms import LoginForm
-from forms import CreateTableForm
-
+from forms import RegistrationForm, LoginForm, CreateTableForm
 
 app = Flask(__name__)
-app.secret_key = "Gxf613UhGRkzAKd47R5daLrUelnlUL4L6AU4z0uu++TNBpdzhAolufHqPQiiEdn34pbE97bmXbN"
+app.secret_key = config.SECRET_KEY
 login_manager = LoginManager(app)
+login_manager.login_view = "home"
 
 DB = DBHelper()
 PH = PasswordHelper()
@@ -40,6 +24,7 @@ def load_user(user_id):
     user_password = DB.get_user(user_id)
     if user_password:
         return User(user_id)
+    return None
 
 
 @app.route("/login", methods=["POST"])
@@ -47,11 +32,11 @@ def login():
     form = LoginForm(request.form)
     if form.validate():
         stored_user = DB.get_user(form.loginemail.data)
-        if stored_user and PH.validate_password(form.loginpassword.data, stored_user['salt'], stored_user['hashed']):
+        if stored_user and PH.validate_password(form.loginpassword.data, stored_user["salt"], stored_user["hashed"]):
             user = User(form.loginemail.data)
             login_user(user, remember=True)
-            return redirect(url_for('account'))
-        form.loginemail.errors.append("Email or password invalid")
+            return redirect(url_for("account"))
+        form.loginemail.errors.append("Invalid email or password.")
     return render_template("home.html", loginform=form, registrationform=RegistrationForm())
 
 
@@ -60,12 +45,14 @@ def register():
     form = RegistrationForm(request.form)
     if form.validate():
         if DB.get_user(form.email.data):
-            form.email.errors.append("Email address already registered")
+            form.email.errors.append("Email address is already registered.")
             return render_template("home.html", loginform=LoginForm(), registrationform=form)
         salt = PH.get_salt()
-        hashed = PH.get_hash(form.password2.data + salt)
+        hashed = PH.get_hash(form.password.data + salt)
         DB.add_user(form.email.data, salt, hashed)
-        return render_template("home.html", loginform=LoginForm(), registrationform=form, onloadmessage="Registration successful. Please log in.")
+        return render_template(
+            "home.html", loginform=LoginForm(), registrationform=form, onloadmessage="Registration successful. Please log in."
+        )
     return render_template("home.html", loginform=LoginForm(), registrationform=form)
 
 
@@ -86,8 +73,9 @@ def dashboard():
     now = datetime.datetime.now()
     requests = DB.get_requests(current_user.get_id())
     for req in requests:
-        deltaseconds = (now - req['time']).seconds
-        req['wait_minutes'] = "{}.{}".format((deltaseconds/60), str(deltaseconds % 60).zfill(2))
+        deltaseconds = (now - req["time"]).seconds
+        minutes, seconds = divmod(deltaseconds, 60)
+        req["wait_minutes"] = f"{minutes}:{str(seconds).zfill(2)}"
     return render_template("dashboard.html", requests=requests)
 
 
@@ -95,8 +83,9 @@ def dashboard():
 @login_required
 def dashboard_resolve():
     request_id = request.args.get("request_id")
-    DB.delete_request(request_id)
-    return redirect(url_for('dashboard'))
+    if request_id:
+        DB.delete_request(request_id)
+    return redirect(url_for("dashboard"))
 
 
 @app.route("/account")
@@ -112,9 +101,9 @@ def account_createtable():
     form = CreateTableForm(request.form)
     if form.validate():
         tableid = DB.add_table(form.tablenumber.data, current_user.get_id())
-        new_url = BH.shorten_url(config.base_url + "newrequest/" + str(tableid))
+        new_url = BH.shorten_url(config.base_url + f"newrequest/{tableid}")
         DB.update_table(tableid, new_url)
-        return redirect(url_for('account'))
+        return redirect(url_for("account"))
     return render_template("account.html", createtableform=form, tables=DB.get_tables(current_user.get_id()))
 
 
@@ -122,16 +111,17 @@ def account_createtable():
 @login_required
 def account_deletetable():
     tableid = request.args.get("tableid")
-    DB.delete_table(tableid)
-    return redirect(url_for('account'))
+    if tableid:
+        DB.delete_table(tableid)
+    return redirect(url_for("account"))
 
 
 @app.route("/newrequest/<tid>")
 def new_request(tid):
     if DB.add_request(tid, datetime.datetime.now()):
-        return "Your request has been logged and a waiter will be with you shortly"
-    return "There is already a request pending for this table. Please be patient, a waiter will be there ASAP"
+        return "Your request has been logged, and a waiter will be with you shortly."
+    return "A request is already pending for this table. Please be patient, a waiter will assist you soon."
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     app.run(debug=True)
